+++ common/server.h
@@ -17,22 +17,22 @@ class Server : public QObject
     Q_OBJECT
 protected:
     std::vector<std::unique_ptr<ServerProtocolHandler>> mClients;
-    std::unordered_map<size_t, std::unique_ptr<ServerGame>> mGames;
+    std::unordered_map<int, std::unique_ptr<ServerGame>> mGames;
 
-    size_t mNextGameId;
+    int mNextGameId;
 
 public:
     Server();
 
     QReadWriteLock mGamesLock;
-    std::unordered_map<size_t, std::unique_ptr<ServerGame>>& games() {
+    std::unordered_map<int, std::unique_ptr<ServerGame>>& games() {
         return mGames;
     }
-    ServerGame* game(size_t id);
+    ServerGame* game(int id);
 
     void createGame(const CommandCreateGame &cmd, ServerProtocolHandler *client);
     void processGameJoinRequest(const CommandJoinGame &cmd, ServerProtocolHandler *client);
 
 protected:
-    size_t nextGameId();
+    int nextGameId();
 };
