+++ common/serverGame.h
@@ -20,28 +20,28 @@ enum class Phase {
 
 class ServerGame
 {
-    size_t mId;
-    size_t mNextPlayerId;
+    int mId;
+    int mNextPlayerId;
     std::string mDescription;
-    std::unordered_map<size_t, std::unique_ptr<ServerPlayer>> mPlayers;
+    std::unordered_map<int, std::unique_ptr<ServerPlayer>> mPlayers;
     Phase mCurrentPhase = Phase::Mulligan;
 
 public:
-    ServerGame(size_t id, std::string description);
+    ServerGame(int id, std::string description);
 
     QMutex mGameMutex;
 
-    std::unordered_map<size_t, std::unique_ptr<ServerPlayer>>& players() {
+    std::unordered_map<int, std::unique_ptr<ServerPlayer>>& players() {
         return mPlayers;
     }
-    ServerPlayer* player(size_t id);
+    ServerPlayer* player(int id);
     void addPlayer(ServerProtocolHandler *client);
-    ServerPlayer* opponentOfPlayer(size_t id) const;
+    ServerPlayer* opponentOfPlayer(int id) const;
     void setStartingPlayer();
     void startGame();
     void endMulligan();
     Phase phase() const { return mCurrentPhase; }
     void setPhase(Phase phase) { mCurrentPhase = phase; }
 
-    void sendPublicEvent(const ::google::protobuf::Message &event, size_t senderId);
+    void sendPublicEvent(const ::google::protobuf::Message &event, int senderId);
 };
